fix: for Android 12 we can have precise location denied but approximate location given
In Android 12 users can now choose to provide only an approximate location. For that reason, the way we were checking the permissions, if the precise location is denied (even if the approximate location is given), the location permission denied error would be returned.

References: https://outsystemsrd.atlassian.net/browse/RMET-823