fix: check if LocationContext is null before using to send error in callback
If the user takes too long to Deny the Location permission in the AddWatch action, a timeout error expired error is sent and the LocationContext becomes null. We need to avoid sending the LOCATION_PERMISSION_DENIED error in this case (it would not be shown anyway, the timeout expired error is shown in this case).

Also moved the code from line 298 inside the else branch. It does not make sense to remove the locationContext in the if case.

References: https://outsystemsrd.atlassian.net/browse/RMET-964